Skyler - Optimized TeleopShootCmd and Updated ShootCmd
Optimized TeleopShootCmd to only calculate speed with Vision Sub if left bumper is not pressed to save processing time. Updated ShootCmd to be optimized the same way as well as making the speed parameter optional, and allowing it so if the user gives ShootCmd a speed parameter, it will instead use that as long as it is greater than or equal to zero. (This should be used to stop the shooter in autonomous mode.)